nfdiv-1158: tidyup state sequence
